* added createdTime to user
Also removed nest() from service layer since it's dumb and that's not how this works. That's not how any of this works!  Basically it was picking up the response method so it was acting liek there were two requests coming in, resulting in a Not Found response after the initial 200